Techincal debts
Typo fixing.
XSD cleaning up.
Fixed event XSD location.
Fixed instance.xsd.
Fixed typo in conquerable clan halls table.

+++ src/main/java/com/l2jserver/datapack/quests/Q00637_ThroughOnceMore/Q00637_ThroughOnceMore.java
@@ -103,7 +103,7 @@ public final class Q00637_ThroughOnceMore extends Quest {
 			return getNoQuestMsg(player);
 		}
 		
-		final byte id = st.getState();
+		final int id = st.getState();
 		if (id == State.CREATED) {
 			if (player.getLevel() > 72) {
 				if (st.hasQuestItems(FADED_MARK)) {
